Added message classes. Partially implemented methods for receiving and handling of incoming msgs, that form the basis for phase1 and phase2 of paxos. Added methods for unicast(), broadcast() and, decide() that evnetually will handled updating log and CLI status

+++ src/bankaccount/AcceptNotificationMessage.java
@@ -4,9 +4,9 @@ import bankaccount.Pair;
 
 public class AcceptNotificationMessage extends Message {
 	private Pair ballotNum;
-	private int value;
+	private double value;
 	
-	public AcceptNotificationMessage(Pair ballotNum, int value) {
+	public AcceptNotificationMessage(Pair ballotNum, double value) {
 		this.ballotNum = ballotNum;
 		this.value = value;
 	}
@@ -15,7 +15,7 @@ public class AcceptNotificationMessage extends Message {
 		return ballotNum;
 	}
 
-	public int getValue() {
+	public double getValue() {
 		return value;
 	}
 }
